requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
attrs==26.1.0
bracex==2.6
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.6
dataclasses-json==0.6.7
deepagents==0.4.12
distro==1.9.0
docstring_parser==0.17.0
faiss-cpu==1.13.2
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.49.1
google-genai==1.70.0
greenlet==3.3.2
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-anthropic==1.4.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.26
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
modal==1.4.1
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
propcache==0.4.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pypdf==6.9.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.49
synchronicity==0.12.1
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tqdm==4.67.3
typer==0.24.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
watchfiles==1.1.1
wcmatch==10.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
